Raster pass materials structured buffer complete

+++ src/Shaders/pixel.hlsl
@@ -14,7 +14,7 @@ struct PixelIn
 cbuffer cbPerObject : register(b0)
 {
     float4x4 gWorld;
-    float matIndex;
+    int matIndex;
 }
